Move timer event logic to TimerModelStateManager
- Timer model states get mutated in one place
- Includes default behavior for tap-on and tap-off
- Includes logic for rest warning and enforcing/de-enforcing a rest period
- Uses delegate (ViewModel) to show alerts/notifications
- Tests are passing
