+++ Hourglass/Hourglass/Views/ContentView.swift
@@ -26,10 +26,8 @@ struct ContentView_Previews: PreviewProvider {
     static var previews: some View {
         let viewModel = ViewModel(dataManager: DataManager.shared,
                                   settingsManager: SettingsManager.shared,
-                                  timerEventForwarder: ProgressTrackingManager.shared,
                                   timerManager: TimerManager.shared,
-                                  userNotificationManager: UserNotificationManager.shared,
-                                  windowCoordinator: NSApplication.shared.delegate as! WindowCoordinator)
+                                  userNotificationManager: UserNotificationManager.shared)
         ContentView(viewModel: viewModel)
             .font(Font.poppins)
     }
